--- src/pages/applications/NewOrEdit/v2/Template/index.less
@@ -1,27 +1,0 @@
-.card {
-  height: 150px;
-  border-width: 1px;
-}
-
-.card-before-clicked {
-  background-color: white;
-}
-
-.card-after-clicked {
-  background-color: #f1fafe;
-  border-color: #00a2c9
-}
-
-.awsui-cards-card-header {
-  font-size: 1.5rem;
-}
-
-.awsui-cards-card-header-inner {
-  width: 90%;
-  color: #0073bb
-}
-
-.radio {
-  float: right;
-}
-
